requirements.txt
apturl==0.5.2
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==40.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.7
django-appconf==1.0.6
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-select2==8.1.2
django-taggit==4.0.0
django-taggit-autosuggest==0.4.1
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
duplicity==0.8.21
factory-boy==3.3.0
Faker==19.13.0
fasteners==0.14.1
filelock==3.10.7
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==6.8.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==10.1.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.2.0
pluggy==0.13.1
protobuf==3.12.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==4.6.4
pytest-django==3.10.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.30.0
requests-oauthlib==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
virtualenv==20.21.0
wadllib==1.3.6
wcwidth==0.2.9
xdg==5
xkit==0.0.0
zipp==3.17.0
